+++ tests/test_container.py
@@ -33,7 +33,18 @@ class DummyContainer(ItemAccessMixin, BaseDataContainer):
 
 def test_init():
     """Tests initialisation of the DummyContainer class"""
-    DummyContainer(name="dummy", data="foo")
+    # Simple init
+    dc = DummyContainer(name="dummy", data="foo")
+
+    # Assert the name is a string
+    assert isinstance(dc.name, str)
+
+    # Check invalid name arguments
+    with pytest.raises(TypeError, match="Name for DummyContainer needs"):
+        DummyContainer(name=123, data="foo")
+
+    with pytest.raises(ValueError, match="Name for DummyContainer cannot "):
+        DummyContainer(name="a/name/with/the/PATH_JOIN_CHAR", data="foo")
 
 
 def test_check_data_mixin():