+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## v0.2 _(unreleased)_
 - #19: Test for multiple Python versions
+- #20: Make it possible to create custom groups during `DataManager` initialisation
 
 
 ## v0.1